changed the type of properties size and alignment to size_t
Minor code clean up

+++ Karma/src/Karma/Core/Object.cpp
@@ -64,8 +64,8 @@ namespace Karma
 		const std::string& Name,
 		UClass*& ReturnClass,
 		/*void(*RegisterNativeFunc)(),*/
-		uint32_t InSize,
-		uint32_t InAlignment,
+		size_t InSize,
+		size_t InAlignment,
 		/*EClassFlags InClassFlags,
 		EClassCastFlags InClassCastFlags,
 		const TCHAR* InConfigName,
